Use driver's default class loader when user-supplied loader fails
This commit introduces the possiblity of loading a class by the driver
default class loader, if the user-supplied class loader is unable to
locate the class.

This is useful in OSGi deployments where the user-supplied loader
may be able to load some classes but not all of them.

Besides, the driver bundle, in OSGi, has a "Dynamic-Import:*" directive
that makes its class loader capable of locating a great number of
classes.